java: rename linear_programming package → optimization (+ MIPCallbackIT)
Per review feedback (chris-maes):

  com.nvidia.cuopt.linear_programming  →  com.nvidia.cuopt.optimization

Reviewer's reasoning: cuOpt covers more than linear programming now
(routing, etc.), so `linear_programming` is a misleading namespace.
The new `com.nvidia.cuopt.optimization` is broad enough to host the
LP/MILP/QP API alongside future siblings (e.g.
`com.nvidia.cuopt.routing`).

Also renames the MIP callback IT class:

  MipCallbackIT  →  MIPCallbackIT

Cascade: 18 main classes + 5 test classes moved from
linear_programming/ to optimization/, and ~10 files outside the
package updated their imports. Full IT suite (4 unit + 6 IT) passes
locally after the rename.

The C-side header path (`cpp/include/cuopt/linear_programming/cuopt_c.h`)
is unchanged — only the Java package was renamed.
