+++ java/cuopt-java/src/main/java/com/nvidia/cuopt/optimization/MIPStats.java
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.nvidia.cuopt.linear_programming;
+package com.nvidia.cuopt.optimization;
 
 /**
  * Solver statistics for MILP solves. Returned by
