+++ java/cuopt-java/src/main/java22/com/nvidia/cuopt/internal/cuOptProviderImpl.java
@@ -8,21 +8,21 @@ import com.nvidia.cuopt.cuOptException;
 import com.nvidia.cuopt.internal.panama.cuOptMIPGetSolutionCallback;
 import com.nvidia.cuopt.internal.panama.cuOptMIPSetSolutionCallback;
 import com.nvidia.cuopt.internal.panama.cuopt_c_h;
-import com.nvidia.cuopt.linear_programming.CType;
-import com.nvidia.cuopt.linear_programming.Constraint;
-import com.nvidia.cuopt.linear_programming.ErrorStatus;
-import com.nvidia.cuopt.linear_programming.LinearExpr;
-import com.nvidia.cuopt.linear_programming.LpStats;
-import com.nvidia.cuopt.linear_programming.MIPGetSolutionCallback;
-import com.nvidia.cuopt.linear_programming.MIPSetSolutionCallback;
-import com.nvidia.cuopt.linear_programming.MIPStats;
-import com.nvidia.cuopt.linear_programming.Problem;
-import com.nvidia.cuopt.linear_programming.QuadraticExpr;
-import com.nvidia.cuopt.linear_programming.SolverMethod;
-import com.nvidia.cuopt.linear_programming.SolverSettings;
-import com.nvidia.cuopt.linear_programming.TerminationStatus;
-import com.nvidia.cuopt.linear_programming.VType;
-import com.nvidia.cuopt.linear_programming.Variable;
+import com.nvidia.cuopt.optimization.CType;
+import com.nvidia.cuopt.optimization.Constraint;
+import com.nvidia.cuopt.optimization.ErrorStatus;
+import com.nvidia.cuopt.optimization.LinearExpr;
+import com.nvidia.cuopt.optimization.LpStats;
+import com.nvidia.cuopt.optimization.MIPGetSolutionCallback;
+import com.nvidia.cuopt.optimization.MIPSetSolutionCallback;
+import com.nvidia.cuopt.optimization.MIPStats;
+import com.nvidia.cuopt.optimization.Problem;
+import com.nvidia.cuopt.optimization.QuadraticExpr;
+import com.nvidia.cuopt.optimization.SolverMethod;
+import com.nvidia.cuopt.optimization.SolverSettings;
+import com.nvidia.cuopt.optimization.TerminationStatus;
+import com.nvidia.cuopt.optimization.VType;
+import com.nvidia.cuopt.optimization.Variable;
 import com.nvidia.cuopt.spi.cuOptProvider;
 import com.nvidia.cuopt.spi.SolveResult;
 import java.lang.foreign.Arena;
