+++ java/cuopt-java/src/test/java/com/nvidia/cuopt/optimization/MilpSolverIT.java
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.nvidia.cuopt.linear_programming;
+package com.nvidia.cuopt.optimization;
 
 import static com.nvidia.cuopt.cuOpt.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
